requirements.txt
# Flask and web dependencies
flask==3.0.0
flask-cors==4.0.0
werkzeug==3.0.1

# Machine Learning (PyTorch)
torch>=2.1.0
torchvision>=0.16.0
pillow>=10.0.0
numpy>=1.26.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
requests==2.31.0

# Development
python-dotenv==1.0.0

# Production server
gunicorn==21.2.0